refactor(dtos): restructure application and job DTOs with enhanced analytics
- Add ApplicationsByDepartment, ApplicationsBySource, and ApplicationsByMonth to ApplicationStatsDto for granular analytics
- Add ReviewedBy field to ApplicationSubmitDto for tracking reviewer information
- Create new JobApplicationSummaryDto for simplified job application listings
- Remove JobUpdateDto from JobResponseDto as it duplicates job creation/update logic
- Expand JobStatsDto with FilledJobs, ArchivedJobs status tracking and department/employment type breakdowns
- Add EmploymentType, ExperienceLevel, and IsRemote filters to JobFilterDto for improved job discovery
- Convert AverageApplicationsPerJob from int to double for precise metrics calculation
- Add OldestJobDate and NewestJobDate tracking and JobsWithNoApplications metric to JobStatsDto
- These changes enable more detailed recruitment analytics and flexible job filtering across the platform

+++ backend/RecruitmentAPI/Repository/Implementations/ApplicationRepository.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using RecruitmentAPI.Data;
+using RecruitmentAPI.DTOs;
 using RecruitmentAPI.Models;
 using RecruitmentAPI.Repository.Interfaces;
 
@@ -278,9 +279,9 @@ namespace RecruitmentAPI.Repository.Implementations
             if (applications.Any(a => a.AI_Score.HasValue))
             {
                 var scores = applications.Where(a => a.AI_Score.HasValue).Select(a => a.AI_Score.Value);
-                stats.AverageAIScore = scores.Average();
-                stats.HighestAIScore = scores.Max();
-                stats.LowestAIScore = scores.Min();
+                stats.AverageAIScore = (decimal)scores.Average();
+                stats.HighestAIScore = (decimal)scores.Max();
+                stats.LowestAIScore = (decimal)scores.Min();
             }
 
             stats.ApplicationsByDepartment = applications
